Remove obsolete Docker modules and related tests
